Split helper files into several partials
