+++ library/src/scripts/content/countStyles.ts
@@ -4,10 +4,9 @@
  * @license GPL-2.0-only
  */
 
-import { absolutePosition, colorOut, debugHelper, unit } from "@library/styles/styleHelpers";
+import { absolutePosition, colorOut, ColorValues, unit } from "@library/styles/styleHelpers";
 import { globalVariables } from "@library/styles/globalStyleVars";
 import { styleFactory, useThemeCache, variableFactory } from "@library/styles/styleUtils";
-import { ColorValues } from "@library/forms/buttonStyles";
 
 export const countVariables = useThemeCache(() => {
     const globalVars = globalVariables();
