+++ library/src/scripts/styles/shadowHelpers.ts
@@ -8,7 +8,7 @@ import { useThemeCache } from "@library/styles/styleUtils";
 import { BorderRadiusProperty } from "csstype";
 import { ColorHelper } from "csx";
 import { TLength } from "typestyle/lib/types";
-import { ColorValues } from "@library/forms/buttonStyles";
+import { ColorValues } from "@library/styles/styleHelpersColors";
 
 export const shadowHelper = useThemeCache(() => {
     const globalVars = globalVariables();
